fix: decouple spotify error logging from next/server
- Create `services/spotifyErrorLogging.server.ts` to handle logging without `next/server` dependency.
- Refactor `services/spotifyApiErrorHandling.server.ts` to use the new logging module.
- Update `services/spotifyPolling.ts` to import from `spotifyErrorLogging.server.ts`, preventing `AsyncLocalStorage` crash in custom server.
